Add static cast to float for M_PI

+++ src/iCubProprioception/src/CADSuperimposer.cpp
@@ -140,7 +140,7 @@ CADSuperimposer::CADSuperimposer(const yarp::os::ConstString& robot, const yarp:
     drawer_ = new SICAD(cad_hand_,
                         cam_width_, cam_height_, cam_fx_, cam_fy_, cam_cx_, cam_cy_,
                         1,
-                        {1.0, 0.0, 0.0, M_PI},
+                        {1.0, 0.0, 0.0, static_cast<float>(M_PI)},
                         shader_path_,
                         false);
